Module Mssql.Row

type t

val compare : t -> t -> int

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val create_exn : month_offset:int -> colnames:string list -> Freetds.Dblib.data list -> t

val to_alist : t -> (string * string) list

to_alist t converts t to a list of (colname, value) pairs

val bignum : t -> string -> Bignum.t option

bignum colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to a bignum or wasn't in the query.

val bignum_exn : t -> string -> Bignum.t

bignum_exn colname similar to bignum but asserts that the column named colname exists and the data is not null.

val bool : t -> string -> bool option

bool colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to a bool or wasn't in the query.

val bool_exn : t -> string -> bool

bool_exn colname similar to bool but asserts that the column named colname exists and the data is not null.

val float : t -> string -> float option

float colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to a float or wasn't in the query.

val float_exn : t -> string -> float

float_exn colname similar to float but asserts that the column named colname exists and the data is not null.

val int : t -> string -> int option

int colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to an int or wasn't in the query.

val int_exn : t -> string -> int

int_exn colname similar to int but asserts that the column named colname exists and the data is not null.

val int32 : t -> string -> int32 option

int32 colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to an int32 or wasn't in the query.

val int32_exn : t -> string -> int32

int32_exn colname similar to int32 but asserts that the column named colname exists and the data is not null.

val int64 : t -> string -> int64 option

int64 colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to an int64 or wasn't in the query.

val int64_exn : t -> string -> int64

int64_exn colname similar to int64 but asserts that the column named colname exists and the data is not null.

val str : t -> string -> string option

str colname returns the value of the given column, or None if the column is null. Throws an exception if the column wasn't in the query.

val str_exn : t -> string -> string

str_exn colname similar to str but asserts that the column named colname exists and the data is not null.

val date : t -> string -> Core.Date.t option

date colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to a bool or wasn't in the query.

val date_exn : t -> string -> Core.Date.t

date_exn colname similar to date but asserts that the column named colname exists and the data is not null.

val datetime : t -> string -> Core.Time.t option

datetime colname returns the value of the given column, or None if the column is null. Throws an exception if the column can't be converted to a bool or wasn't in the query.

val datetime_exn : t -> string -> Core.Time.t

datetime_exn colname similar to datetime but asserts that the column named colname exists and the data is not null.